requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.0.76
firebase==4.0.1
firebase-admin==6.2.0